requirements-gpu.txt
bleach==1.5.0
boto==2.49.0
boto3==1.9.60
botocore==1.12.60
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
Cython==0.29.1
docutils==0.14
emoji==0.5.1
emot==2.1
enum34==1.1.6
fasttext==0.8.3
future==0.17.1
gensim==3.6.0
html5lib==0.9999999
idna==2.7
jmespath==0.9.3
Markdown==3.0.1
nltk==3.4
numpy==1.15.4
pandas==0.23.4
protobuf==3.6.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
s3transfer==0.1.13
scikit-learn==0.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
smart-open==1.7.1
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0
urllib3==1.24.1
Werkzeug==0.14.1
